requirements.txt
anyio==3.7.1
cachetools==5.3.1
certifi==2023.7.22
click==8.1.6
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
python-telegram-bot==20.4
sniffio==1.3.0
starlette==0.31.0
uvicorn==0.23.1
